# 2018 11 21 performance improvements