--- a/core/src/main/scala/org/bitcoins/core/script/arithmetic/ArithmeticOperations.scala
+++ b/core/src/main/scala/org/bitcoins/core/script/arithmetic/ArithmeticOperations.scala
@@ -8,140 +8,140 @@ sealed trait ArithmeticOperation extends ScriptOperation
 
 /** 1 is added to the input. */
 case object OP_1ADD extends ArithmeticOperation {
-  override def opCode = 139
+  override val opCode: Int = 139
 }
 
 /** 1 is subtracted from the input. */
 case object OP_1SUB extends ArithmeticOperation {
-  override def opCode = 140
+  override val opCode: Int = 140
 }
 
 /** The sign of the input is flipped. */
 case object OP_NEGATE extends ArithmeticOperation {
-  override def opCode = 143
+  override val opCode: Int = 143
 }
 
 /** The input is made positive. */
 case object OP_ABS extends ArithmeticOperation {
-  override def opCode = 144
+  override val opCode: Int = 144
 }
 
 /** If the input is 0 or 1, it is flipped. Otherwise the output will be 0. */
 
 case object OP_NOT extends ArithmeticOperation {
-  override def opCode = 145
+  override val opCode: Int = 145
 }
 
 /** Returns 0 if the input is 0. 1 otherwise. */
 case object OP_0NOTEQUAL extends ArithmeticOperation {
-  override def opCode = 146
+  override val opCode: Int = 146
 }
 
 /** a is added to b. */
 case object OP_ADD extends ArithmeticOperation {
-  override def opCode = 147
+  override val opCode: Int = 147
 }
 
 /** b is subtracted from a. */
 case object OP_SUB extends ArithmeticOperation {
-  override def opCode = 148
+  override val opCode: Int = 148
 }
 
 /** If both a and b are not 0, the output is 1. Otherwise 0. */
 case object OP_BOOLAND extends ArithmeticOperation {
-  override def opCode = 154
+  override val opCode: Int = 154
 }
 
 /** If a or b is not 0, the output is 1. Otherwise 0. */
 case object OP_BOOLOR extends ArithmeticOperation {
-  override def opCode = 155
+  override val opCode: Int = 155
 }
 
 /** Returns 1 if the numbers are equal, 0 otherwise. */
 case object OP_NUMEQUAL extends ArithmeticOperation {
-  override def opCode = 156
+  override val opCode: Int = 156
 }
 
 /** Same as OP_NUMEQUAL, but runs OP_VERIFY afterward. */
 case object OP_NUMEQUALVERIFY extends ArithmeticOperation {
-  override def opCode = 157
+  override val opCode: Int = 157
 }
 
 /** Returns 1 if the numbers are not equal, 0 otherwise. */
 case object OP_NUMNOTEQUAL extends ArithmeticOperation {
-  override def opCode = 158
+  override val opCode: Int = 158
 }
 
 /** Returns 1 if a is less than b, 0 otherwise. */
 case object OP_LESSTHAN extends ArithmeticOperation {
-  override def opCode = 159
+  override val opCode: Int = 159
 }
 
 /** Returns 1 if a is greater than b, 0 otherwise. */
 case object OP_GREATERTHAN extends ArithmeticOperation {
-  override def opCode = 160
+  override val opCode: Int = 160
 }
 
 /** Returns 1 if a is less than or equal to b, 0 otherwise. */
 case object OP_LESSTHANOREQUAL extends ArithmeticOperation {
-  override def opCode = 161
+  override val opCode: Int = 161
 }
 
 /** 	Returns 1 if a is greater than or equal to b, 0 otherwise. */
 case object OP_GREATERTHANOREQUAL extends ArithmeticOperation {
-  override def opCode = 162
+  override val opCode: Int = 162
 }
 
 /** Returns the smaller of a and b. */
 case object OP_MIN extends ArithmeticOperation {
-  override def opCode = 163
+  override val opCode: Int = 163
 }
 
 /** Returns the larger of a and b. */
 case object OP_MAX extends ArithmeticOperation {
-  override def opCode = 164
+  override val opCode: Int = 164
 }
 
 /** Returns 1 if x is within the specified range (left-inclusive), 0 otherwise. */
 case object OP_WITHIN extends ArithmeticOperation {
-  override def opCode = 165
+  override val opCode: Int = 165
 }
 
 //currently disabled operations
 
 /** The input is multiplied by 2. disabled. */
 case object OP_2MUL extends ArithmeticOperation {
-  override def opCode = 141
+  override val opCode: Int = 141
 }
 
 /** The input is divided by 2. disabled. */
 case object OP_2DIV extends ArithmeticOperation {
-  override def opCode = 142
+  override val opCode: Int = 142
 }
 
 /** a is multiplied by b. disabled. */
 case object OP_MUL extends ArithmeticOperation {
-  override def opCode = 149
+  override val opCode: Int = 149
 }
 
 /** a is divided by b. disabled. */
 case object OP_DIV extends ArithmeticOperation {
-  override def opCode = 150
+  override val opCode: Int = 150
 }
 
 /** Returns the remainder after dividing a by b. disabled. */
 case object OP_MOD extends ArithmeticOperation {
-  override def opCode = 151
+  override val opCode: Int = 151
 }
 
 /** Shifts a left b bits, preserving sign. disabled. */
 case object OP_LSHIFT extends ArithmeticOperation {
-  override def opCode = 152
+  override val opCode: Int = 152
 }
 
 /** Shifts a right b bits, preserving sign. disabled. */
 case object OP_RSHIFT extends ArithmeticOperation {
-  override def opCode = 153
+  override val opCode: Int = 153
 }
 
 object ArithmeticOperation extends ScriptOperationFactory[ArithmeticOperation] {