--- a/core/src/main/scala/org/bitcoins/core/script/arithmetic/ArithmeticOperations.scala
+++ b/core/src/main/scala/org/bitcoins/core/script/arithmetic/ArithmeticOperations.scala
@@ -4,7 +4,7 @@ import org.bitcoins.core.script.ScriptOperationFactory
 import org.bitcoins.core.script.constant.ScriptOperation
 
 /** Created by chris on 1/6/16. */
-sealed trait ArithmeticOperation extends ScriptOperation
+sealed abstract class ArithmeticOperation extends ScriptOperation
 
 /** 1 is added to the input. */
 case object OP_1ADD extends ArithmeticOperation {
@@ -145,7 +145,7 @@ case object OP_RSHIFT extends ArithmeticOperation {
 }
 
 object ArithmeticOperation extends ScriptOperationFactory[ArithmeticOperation] {
-  override def operations = Seq(OP_0NOTEQUAL, OP_1ADD, OP_1SUB, OP_ABS, OP_ADD, OP_BOOLAND, OP_BOOLOR,
+  override val operations = Seq(OP_0NOTEQUAL, OP_1ADD, OP_1SUB, OP_ABS, OP_ADD, OP_BOOLAND, OP_BOOLOR,
     OP_GREATERTHAN, OP_GREATERTHANOREQUAL, OP_LESSTHAN, OP_LESSTHANOREQUAL, OP_MAX, OP_MIN, OP_NEGATE,
     OP_NEGATE, OP_NOT, OP_NUMEQUAL, OP_NUMEQUALVERIFY, OP_NUMNOTEQUAL, OP_SUB, OP_WITHIN,
     OP_2MUL, OP_2DIV, OP_MUL, OP_DIV, OP_MOD, OP_LSHIFT, OP_RSHIFT)