--- a/core/src/main/scala/org/bitcoins/core/script/bitwise/BitwiseOperations.scala
+++ b/core/src/main/scala/org/bitcoins/core/script/bitwise/BitwiseOperations.scala
@@ -10,32 +10,32 @@ sealed trait BitwiseOperation extends ScriptOperation
 
 /** Returns 1 if the inputs are exactly equal, 0 otherwise. */
 case object OP_EQUAL extends BitwiseOperation {
-  override def opCode = 135
+  override val opCode: Int = 135
 }
 
 /** Same as [[OP_EQUAL]], but runs [[org.bitcoins.core.script.control.OP_VERIFY]] afterward. */
 case object OP_EQUALVERIFY extends BitwiseOperation {
-  override def opCode = 136
+  override val opCode: Int = 136
 }
 
 /** Flips all of the bits in the input. disabled. */
 case object OP_INVERT extends BitwiseOperation {
-  override def opCode = 131
+  override val opCode: Int = 131
 }
 
 /** Boolean and between each bit in the inputs. disabled. */
 case object OP_AND extends BitwiseOperation {
-  override def opCode = 132
+  override val opCode: Int = 132
 }
 
 /** Boolean or between each bit in the inputs. disabled. */
 case object OP_OR extends BitwiseOperation {
-  override def opCode = 133
+  override val opCode: Int = 133
 }
 
 /** Boolean exclusive or between each bit in the inputs. disabled. */
 case object OP_XOR extends BitwiseOperation {
-  override def opCode = 134
+  override val opCode: Int = 134
 }
 
 object BitwiseOperation extends ScriptOperationFactory[BitwiseOperation] {