--- a/core/src/main/scala/org/bitcoins/core/script/constant/Constants.scala
+++ b/core/src/main/scala/org/bitcoins/core/script/constant/Constants.scala
@@ -21,7 +21,7 @@ sealed trait ScriptToken extends NetworkElement {
   def bytes: ByteVector
 
   /** The conversion from the byte representation of a [[ScriptToken]] to a number. */
-  def toLong = ScriptNumberUtil.toLong(hex)
+  def toLong: Long = ScriptNumberUtil.toLong(bytes)
 }
 
 /**
@@ -31,7 +31,9 @@ sealed trait ScriptToken extends NetworkElement {
 trait ScriptOperation extends ScriptToken {
   def opCode: Int
 
-  def bytes: ByteVector = ByteVector.fromByte(opCode.toByte)
+  override def bytes: ByteVector = ByteVector.fromByte(toByte)
+
+  def toByte: Byte = opCode.toByte
 }
 
 /** A constant in the Script language for instance as String or a number. */
@@ -88,7 +90,7 @@ sealed abstract class ScriptNumber extends ScriptConstant {
   override def toLong = underlying
 
   /** The underlying number of the [[ScriptNumber]]. */
-  protected def underlying: Long
+  protected val underlying: Long
 }
 
 object ScriptNumber extends Factory[ScriptNumber] {
@@ -146,7 +148,7 @@ object ScriptNumber extends Factory[ScriptNumber] {
 
 /** The next byte contains the number of bytes to be pushed onto the stack. */
 case object OP_PUSHDATA1 extends ScriptOperation {
-  override def opCode = 76
+  override val opCode: Int = 76
 
   /** The maximum amount of bytes OP_PUSHDATA1 can push onto the stack. */
   def max = 255
@@ -154,7 +156,7 @@ case object OP_PUSHDATA1 extends ScriptOperation {
 
 /** The next two bytes contain the number of bytes to be pushed onto the stack. */
 case object OP_PUSHDATA2 extends ScriptOperation {
-  override def opCode = 77
+  override val opCode: Int = 77
 
   /** The max amount of data that OP_PUSHDATA2 can push onto the stack. */
   def max = 65535
@@ -162,7 +164,7 @@ case object OP_PUSHDATA2 extends ScriptOperation {
 
 /** The next four bytes contain the number of bytes to be pushed onto the stack. */
 case object OP_PUSHDATA4 extends ScriptOperation {
-  override def opCode = 78
+  override val opCode: Int = 78
 
   /** The maximum amount of data that OP_PUSHDATA4 can be push on the stack. */
   def max = 4294967295L
@@ -178,148 +180,148 @@ sealed abstract class ScriptNumberOperation extends ScriptNumber with ScriptOper
 
 /** An empty array of bytes is pushed onto the stack. (This is not a no-op: an item is added to the stack.) */
 case object OP_0 extends ScriptNumberOperation {
-  override def opCode = 0
+  override val opCode: Int = 0
 
-  override def hex = "00"
+  override val hex: String = "00"
 
-  override def underlying = 0
+  override val underlying: Long = 0
 }
 
 /** An empty array of bytes is pushed onto the stack. (This is not a no-op: an item is added to the stack.) */
 case object OP_FALSE extends ScriptNumberOperation {
-  override def opCode = OP_0.opCode
+  override val opCode = OP_0.opCode
 
-  override def hex = OP_0.hex
+  override val hex = OP_0.hex
 
-  override def underlying = OP_0.underlying
+  override val underlying = OP_0.underlying
 
-  override def bytes = OP_0.bytes
+  override val bytes = OP_0.bytes
 }
 
 /** The number 1 is pushed onto the stack. */
 case object OP_TRUE extends ScriptNumberOperation {
-  override def opCode = 81
+  override val opCode = 81
 
-  override def underlying = 1
+  override val underlying: Long = 1
 }
 
 /** The number -1 is pushed onto the stack. */
 case object OP_1NEGATE extends ScriptNumberOperation {
-  override def opCode = 79
+  override val opCode: Int = 79
 
-  override def underlying = -1
+  override val underlying: Long = -1
 }
 
 /** The number 1 is pushed onto the stack. */
 case object OP_1 extends ScriptNumberOperation {
-  override def opCode = OP_TRUE.opCode
+  override val opCode: Int = OP_TRUE.opCode
 
-  override def underlying = OP_TRUE.underlying
+  override val underlying: Long = OP_TRUE.underlying
 }
 
 /** The number 2 is pushed onto the stack. */
 case object OP_2 extends ScriptNumberOperation {
-  override def opCode = 82
+  override val opCode: Int = 82
 
-  override def underlying = 2
+  override val underlying: Long = 2
 }
 
 /** The number 3 is pushed onto the stack. */
 case object OP_3 extends ScriptNumberOperation {
-  override def opCode = 83
+  override val opCode: Int = 83
 
-  override def underlying = 3
+  override val underlying: Long = 3
 }
 
 /** The number 4 is pushed onto the stack. */
 case object OP_4 extends ScriptNumberOperation {
-  override def opCode = 84
+  override val opCode: Int = 84
 
-  override def underlying = 4
+  override val underlying: Long = 4
 }
 
 /** The number 5 is pushed onto the stack. */
 case object OP_5 extends ScriptNumberOperation {
-  override def opCode = 85
+  override val opCode: Int = 85
 
-  override def underlying = 5
+  override val underlying: Long = 5
 }
 
 /** The number 6 is pushed onto the stack. */
 case object OP_6 extends ScriptNumberOperation {
-  override def opCode = 86
+  override val opCode: Int = 86
 
-  override def underlying = 6
+  override val underlying: Long = 6
 }
 
 /** The number 7 is pushed onto the stack. */
 case object OP_7 extends ScriptNumberOperation {
-  override def opCode = 87
+  override val opCode: Int = 87
 
-  override def underlying = 7
+  override val underlying: Long = 7
 }
 
 /** The number 8 is pushed onto the stack. */
 case object OP_8 extends ScriptNumberOperation {
-  override def opCode = 88
+  override val opCode: Int = 88
 
-  override def underlying = 8
+  override val underlying: Long = 8
 }
 
 /** The number 9 is pushed onto the stack. */
 case object OP_9 extends ScriptNumberOperation {
-  override def opCode = 89
+  override val opCode: Int = 89
 
-  override def underlying = 9
+  override val underlying: Long = 9
 }
 
 /** The number 10 is pushed onto the stack. */
 case object OP_10 extends ScriptNumberOperation {
-  override def opCode = 90
+  override val opCode: Int = 90
 
-  override def underlying = 10
+  override val underlying: Long = 10
 }
 
 /** The number 11 is pushed onto the stack. */
 case object OP_11 extends ScriptNumberOperation {
-  override def opCode = 91
+  override val opCode: Int = 91
 
-  override def underlying = 11
+  override val underlying: Long = 11
 }
 
 /** The number 12 is pushed onto the stack. */
 case object OP_12 extends ScriptNumberOperation {
-  override def opCode = 92
+  override val opCode: Int = 92
 
-  override def underlying = 12
+  override val underlying: Long = 12
 }
 
 /** The number 13 is pushed onto the stack. */
 case object OP_13 extends ScriptNumberOperation {
-  override def opCode = 93
+  override val opCode: Int = 93
 
-  override def underlying = 13
+  override val underlying: Long = 13
 }
 
 /** The number 14 is pushed onto the stack. */
 case object OP_14 extends ScriptNumberOperation {
-  override def opCode = 94
+  override val opCode: Int = 94
 
-  override def underlying = 14
+  override val underlying: Long = 14
 }
 
 /** The number 15 is pushed onto the stack. */
 case object OP_15 extends ScriptNumberOperation {
-  override def opCode = 95
+  override val opCode: Int = 95
 
-  override def underlying = 15
+  override val underlying: Long = 15
 }
 
 /** The number 16 is pushed onto the stack. */
 case object OP_16 extends ScriptNumberOperation {
-  override def opCode = 96
+  override val opCode: Int = 96
 
-  override def underlying = 16
+  override val underlying: Long = 16
 }
 
 object ScriptNumberOperation extends ScriptOperationFactory[ScriptNumberOperation] {
@@ -327,7 +329,7 @@ object ScriptNumberOperation extends ScriptOperationFactory[ScriptNumberOperatio
   /** Finds the [[ScriptNumberOperation]] based on the given integer. */
   def fromNumber(underlying: Long): Option[ScriptNumberOperation] = operations.find(_.underlying == underlying)
 
-  def operations = Seq(OP_0, OP_1, OP_1NEGATE, OP_2, OP_3, OP_4, OP_5, OP_6, OP_7, OP_8, OP_9, OP_10, OP_11, OP_12, OP_13, OP_14, OP_15, OP_16)
+  val operations = Seq(OP_0, OP_1, OP_1NEGATE, OP_2, OP_3, OP_4, OP_5, OP_6, OP_7, OP_8, OP_9, OP_10, OP_11, OP_12, OP_13, OP_14, OP_15, OP_16)
 
 }