… bench/src/main/scala/BlockBench.scala
… core-test/src/test/scala/org/bitcoins/core/script/ScriptOperationFactoryTest.scala
… core-test/src/test/scala/org/bitcoins/core/script/constant/ConstantsTest.scala
… core-test/src/test/scala/org/bitcoins/core/script/splice/SpliceOperationFactoryTest.scala
… core/src/main/scala/org/bitcoins/core/script/reserved/ReservedOperations.scala
… core/src/main/scala/org/bitcoins/core/script/stack/StackOperations.scala