This seems to improve performance by roughly 2x with rudimentary bench markets 

```
2018-11-21 16:28:52.323 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 1593ms
2018-11-21 16:28:53.052 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 712ms
2018-11-21 16:28:53.729 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 648ms
2018-11-21 16:28:54.072 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 328ms
2018-11-21 16:28:54.459 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 372ms
2018-11-21 16:28:54.793 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 320ms
2018-11-21 16:28:55.200 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 394ms
2018-11-21 16:28:55.528 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 315ms
2018-11-21 16:28:56.002 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 460ms
2018-11-21 16:28:56.407 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 388ms

Flat profile of 5.97 secs (533 total ticks): main

  Interpreted + native   Method                        
  0.9%     4  +     1    java.lang.ClassLoader.defineClass1
  0.6%     3  +     0    java.lang.Integer.toString
  0.6%     3  +     0    org.bitcoins.core.util.BitcoinScriptUtil$class.castToBool
  0.6%     3  +     0    java.math.BigInteger.toByteArray
  0.4%     0  +     2    java.util.zip.ZipFile.getEntryTime
  0.4%     0  +     2    org.bitcoin.NativeSecp256k1.secp256k1_ec_pubkey_decompress
  0.4%     2  +     0    org.bitcoins.core.script.reserved.UndefinedOP_NOP.bytes
  0.4%     2  +     0    org.bitcoins.core.script.constant.StackPushOperationFactory$class.operations
  0.2%     0  +     1    java.lang.Class.isInterface
  0.2%     0  +     1    org.bitcoin.Secp256k1Context.secp256k1_init_context
  0.2%     1  +     0    java.util.zip.Inflater.reset
  0.2%     0  +     1    java.security.AccessController.doPrivileged
  0.2%     1  +     0    org.bitcoins.core.protocol.script.ScriptPubKey$.fromAsm
  0.2%     1  +     0    org.bitcoins.core.number.Int64$$anonfun$fromBytes$4.<init>
  0.2%     1  +     0    org.bitcoins.core.script.stack.OP_3DUP$.bytes
  0.2%     1  +     0    org.bitcoins.core.script.stack.OP_2ROT$.bytes
  0.2%     1  +     0    org.bitcoins.core.script.stack.OP_SWAP$.bytes
  0.2%     1  +     0    org.bitcoins.core.script.stack.OP_TUCK$.bytes
  0.2%     1  +     0    ch.qos.logback.core.subst.Parser.makeNewLiteralNode
  0.2%     1  +     0    org.bitcoins.core.crypto.ECPublicKey$.fromBytes
  0.2%     1  +     0    org.bitcoins.core.protocol.script.P2PKHScriptSignature$$anonfun$fromAsm$4.apply
  0.2%     1  +     0    org.bitcoins.core.crypto.DoubleSha256Digest$.fromBytes
  0.2%     1  +     0    scodec.bits.ByteVector.toArray
  0.2%     1  +     0    scala.LowPriorityImplicits.wrapByteArray
  0.2%     1  +     0    org.bitcoins.core.script.stack.OP_IFDUP$.<init>
 12.8%    59  +     9    Total interpreted (including elided)

     Compiled + native   Method                        
  3.2%    13  +     4    scodec.bits.ByteVector$.fromHexInternal
  2.3%    12  +     0    scala.collection.IndexedSeqOptimized$class.foreach
  2.1%    11  +     0    org.bitcoins.core.serializers.script.ScriptParser.loop$2
  1.9%    10  +     0    org.bitcoins.core.util.BitcoinSUtil$class.encodeHex
  1.7%     9  +     0    scodec.bits.ByteVector.go$5
  1.7%     9  +     0    java.math.BigInteger.<init>
  1.5%     7  +     1    scala.collection.TraversableOnce$$anonfun$addString$1.apply
  1.3%     7  +     0    scodec.bits.ByteVector$Buffer.$plus$plus
  1.1%     6  +     0    org.bitcoins.core.protocol.script.MultiSignatureScriptPubKey$.isMultiSignatureScriptPubKey
  0.9%     5  +     0    scala.collection.AbstractIterable.foreach
  0.9%     0  +     5    scala.reflect.ManifestFactory$$anon$6.newArray
  0.9%     5  +     0    org.bitcoins.core.script.constant.ScriptNumberOperation.toByte
  0.9%     5  +     0    org.bitcoins.core.util.BitcoinSUtil$$anonfun$toByteVector$1.apply
  0.9%     5  +     0    scala.collection.TraversableLike$class.$plus$plus
  0.8%     4  +     0    org.bitcoins.core.protocol.script.MultiSignatureScriptPubKey$.org$bitcoins$core$protocol$script$MultiSignatureScriptPubKey$$isValidPubKeyNumber
  0.8%     3  +     1    org.bitcoins.core.serializers.RawSerializerHelper.writeCmpctSizeUInt
  0.8%     4  +     0    org.bitcoins.core.protocol.CompactSizeUInt.hex
  0.6%     3  +     0    org.bitcoins.core.script.constant.ScriptNumberOperation.bytes
  0.6%     3  +     0    scala.collection.mutable.StringBuilder.appendAll
  0.6%     2  +     1    scala.collection.immutable.List.foreach
  0.6%     2  +     1    org.bitcoins.core.script.constant.ScriptNumberUtil$class.toLong
  0.4%     2  +     0    scala.collection.immutable.List.find
  0.4%     2  +     0    scodec.bits.ByteVector$Buffer.getImpl
  0.4%     2  +     0    scala.runtime.ScalaRunTime$._toString
  0.4%     2  +     0    org.bitcoins.core.protocol.script.ScriptFactory$class.buildScript
 40.0%   176  +    37    Total compiled (including elided)

         Stub + native   Method                        
 27.8%     0  +   148    org.bitcoin.NativeSecp256k1.secp256k1_ec_pubkey_decompress
  6.0%     0  +    32    java.lang.Throwable.fillInStackTrace
  3.8%     0  +    20    java.lang.reflect.Array.newArray
  3.4%     0  +    18    java.lang.Object.clone
  2.3%     0  +    12    sun.misc.Unsafe.copyMemory
  1.9%     0  +    10    java.util.zip.Inflater.inflateBytes
  1.1%     5  +     1    java.lang.ClassLoader.defineClass1
  0.2%     0  +     1    java.lang.Class.isInterface
  0.2%     0  +     1    java.security.AccessController.doPrivileged
  0.2%     0  +     1    java.lang.ClassLoader.findBootstrapClass
  0.2%     0  +     1    java.lang.ClassLoader.findLoadedClass0
  0.2%     0  +     1    java.lang.System.arraycopy
  0.2%     0  +     1    java.util.zip.ZipFile.getEntry
 47.3%     5  +   247    Total stub


Global summary of 5.99 seconds:
100.0%   542             Received ticks
  1.3%     7             Received GC ticks
142.8%   774             Compilation

```