@Christewart 2018-11-22
Latest profile on c4b665be1c387b0930b4a5d12bb8943c4775844e

```
2018-11-22 11:33:41.103 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 995ms
2018-11-22 11:33:41.534 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 416ms
2018-11-22 11:33:42.035 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 485ms
2018-11-22 11:33:42.333 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 284ms
2018-11-22 11:33:42.728 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 383ms
2018-11-22 11:33:43.003 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 263ms
2018-11-22 11:33:43.267 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 252ms
2018-11-22 11:33:43.577 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 295ms
2018-11-22 11:33:43.846 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 257ms
2018-11-22 11:33:44.179 INFO  BlockBench$.timeBlockParsing(14) - Elapsed time: 320ms

Flat profile of 4.35 secs (397 total ticks): main

  Interpreted + native   Method                        
  1.3%     5  +     0    scala.collection.GenIterable$class.$init$
  1.3%     5  +     0    java.lang.ClassLoader.defineClass1
  0.3%     0  +     1    org.bitcoin.Secp256k1Context.secp256k1_init_context
  0.3%     1  +     0    scodec.bits.ByteVector.tail
  0.3%     0  +     1    java.lang.Throwable.fillInStackTrace
  0.3%     1  +     0    org.bitcoins.core.protocol.script.ScriptPubKey$.apply
  0.3%     1  +     0    scala.collection.immutable.Range.isEmpty
  0.3%     1  +     0    org.bitcoins.core.script.splice.OP_CAT$.opCode
  0.3%     0  +     1    sun.misc.Unsafe.copyMemory
  0.3%     1  +     0    java.lang.Character.isWhitespace
  0.3%     1  +     0    scala.collection.immutable.Range$.apply
  0.3%     1  +     0    scala.runtime.ScalaRunTime$$anon$1.<init>
  0.3%     1  +     0    scala.App$class.$init$
  0.3%     1  +     0    org.bitcoins.core.protocol.script.P2WPKHWitnessV0$.apply
  0.3%     1  +     0    scala.collection.immutable.VectorPointer$class.copyRange
  0.3%     1  +     0    org.bitcoins.core.protocol.script.P2WPKHWitnessSPKV0$.isValid
  0.3%     1  +     0    scodec.bits.ByteVector.bufferBy
  0.3%     1  +     0    org.bitcoins.core.protocol.script.ScriptFactory$class.buildScript
  0.3%     1  +     0    org.bitcoins.core.protocol.script.P2WPKHWitnessSPKV0$.fromAsm
  0.3%     0  +     1    ch.qos.logback.core.joran.GenericConfigurator.doConfigure
  0.3%     1  +     0    java.util.zip.ZipInputStream.<init>
  0.3%     1  +     0    sun.reflect.generics.parser.SignatureParser.parseClassTypeSignature
  0.3%     0  +     1    ch.qos.logback.classic.LoggerContext.<init>
  0.3%     1  +     0    org.bitcoins.core.serializers.script.RawScriptWitnessParser.loop$1
  0.3%     1  +     0    org.bitcoins.core.serializers.transaction.RawWitnessTransactionParser.write
 10.6%    37  +     5    Total interpreted (including elided)

     Compiled + native   Method                        
  3.3%    11  +     2    scodec.bits.ByteVector$.fromHexInternal
  2.3%     9  +     0    scala.collection.IndexedSeqOptimized$class.foreach
  2.3%     9  +     0    org.bitcoins.core.serializers.script.ScriptParser.loop$2
  1.5%     0  +     6    scala.reflect.ManifestFactory$$anon$6.newArray
  1.5%     6  +     0    org.bitcoins.core.script.constant.ScriptNumberOperation.toByte
  1.3%     5  +     0    org.bitcoins.core.script.arithmetic.ArithmeticOperation.toByte
  1.3%     5  +     0    scala.collection.TraversableOnce$$anonfun$addString$1.apply
  1.0%     4  +     0    org.bitcoins.core.protocol.script.MultiSignatureScriptSignature$$anonfun$11.apply
  1.0%     4  +     0    org.bitcoins.core.number.UInt64.encodeHex
  1.0%     4  +     0    scodec.bits.ByteVector.toString
  1.0%     4  +     0    scodec.bits.ByteVector.go$5
  1.0%     3  +     1    java.math.BigInteger.<init>
  0.8%     3  +     0    org.bitcoins.core.serializers.transaction.RawTransactionInputParser.read
  0.8%     3  +     0    org.bitcoins.core.protocol.script.MultiSignatureScriptPubKey$.isMultiSignatureScriptPubKey
  0.5%     2  +     0    ExceptionBlob
  0.5%     2  +     0    scala.collection.AbstractIterable.foreach
  0.5%     2  +     0    org.bitcoins.core.number.UInt32.hex
  0.5%     1  +     1    scala.Predef$.require
  0.5%     2  +     0    org.bitcoins.core.number.UInt32$.fromBytes
  0.5%     2  +     0    scala.collection.TraversableLike$class.$plus$plus
  0.5%     1  +     1    scala.collection.IndexedSeqOptimized$class.zipWithIndex
  0.5%     2  +     0    org.bitcoins.core.serializers.transaction.RawBaseTransactionParser.write
  0.5%     2  +     0    org.bitcoins.core.util.BitcoinSUtil$class.encodeHex
  0.5%     2  +     0    org.bitcoins.core.protocol.script.ScriptPubKey$.fromAsm
  0.3%     1  +     0    scala.collection.immutable.List.find
 32.2%   107  +    21    Total compiled (including elided)

         Stub + native   Method                        
 35.0%     0  +   139    org.bitcoin.NativeSecp256k1.secp256k1_ec_pubkey_decompress
  5.8%     0  +    23    java.lang.Object.clone
  4.5%     0  +    18    java.lang.Throwable.fillInStackTrace
  3.3%     0  +    13    java.util.zip.Inflater.inflateBytes
  2.8%     0  +    11    sun.misc.Unsafe.copyMemory
  2.5%     0  +    10    java.lang.reflect.Array.newArray
  1.0%     3  +     1    java.lang.ClassLoader.defineClass1
  0.8%     0  +     3    java.lang.ClassLoader.findBootstrapClass
  0.8%     0  +     3    java.util.zip.ZipFile.getEntry
  0.3%     0  +     1    java.util.zip.Inflater.reset
  0.3%     0  +     1    java.util.zip.ZipFile.read
 56.9%     3  +   223    Total stub

  Thread-local ticks:
  0.3%     1             Class loader


Global summary of 4.37 seconds:
100.0%   408             Received ticks
  2.2%     9             Received GC ticks
135.0%   551             Compilation
  0.2%     1             Class loader
```