@torkelrogstad 2018-11-23
Very nice! Do you have any idea if it is the `val`/`def`s that did it, or the removal of the serializer loop?