@Christewart 2018-11-23
@torkelrogstad It was mostly the changing from `def` -> `val`. This was a surprising performance increase but i guess it makes sense since by definition `case object`s should never change -- so we should be using `val`s that don't have to get re-evaluated every single time it is executed. 

The removal of the loop was a notable performance increase, but secondary to `def` -> `val`