@torkelrogstad 2018-11-23
Interestring! And very impressive gains.